refactor: move common features to LanguageContext
