+++ src/main/java/io/github/handofgod94/grammar/ElementContext.java
@@ -2,9 +2,6 @@ package io.github.handofgod94.grammar;
 
 import io.github.handofgod94.grammar.visitors.Visitor;
 import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-import org.eclipse.lsp4j.CompletionList;
 import org.eclipse.lsp4j.Position;
 
 public class ElementContext extends LanguageContext implements GrammarElement {
@@ -15,16 +12,8 @@ public class ElementContext extends LanguageContext implements GrammarElement {
   public static final int PUNCTUATION_SEPARATOR_NAMESPACE_XML = 4;
   public static final int ENTITY_NAME_TAG_LOCALNAME_XML = 5;
 
-  private List<CompletionList> completionList;
-  private boolean isMatched = false;
-  private boolean isInScope = false;
-  private Map<Integer, String> capturedGroups;
-  private int matchStart = -1;
-  private int matchEnd = -1;
-
   private final String begRegEx = "(<\\/?)(?:([-\\w\\.]+)((:)))?([-\\w\\.:]+)";
   private final String endRegEx = "(\\/?>)";
-  private final List<Integer> groupsToMatch;
 
   ElementContext(String line, Position position) {
     super(line, position);
@@ -43,60 +32,8 @@ public class ElementContext extends LanguageContext implements GrammarElement {
     visitor.visit(this);
   }
 
-  public List<CompletionList> getCompletionList() {
-    return completionList;
-  }
-
-  public void setCompletionList(List<CompletionList> completionList) {
-    this.completionList = completionList;
-  }
-
-  public void setCapturedGroups(Map<Integer, String> capturedGroups) {
-    this.capturedGroups = capturedGroups;
-  }
-
-  public Map<Integer, String> getCapturedGroups() {
-    return capturedGroups;
-  }
-
-  public boolean isMatched() {
-    return isMatched;
-  }
-
-  public boolean isInScope() {
-    return isInScope;
-  }
-
-  public void setInScope(boolean inScope) {
-    isInScope = inScope;
-  }
-
-  public int getMatchStart() {
-    return matchStart;
-  }
-
-  public void setMatchStart(int matchStart) {
-    this.matchStart = matchStart;
-  }
-
-  public int getMatchEnd() {
-    return matchEnd;
-  }
-
-  public void setMatchEnd(int matchEnd) {
-    this.matchEnd = matchEnd;
-  }
-
-  public void setMatched(boolean matched) {
-    isMatched = matched;
-  }
-
   public String getBegRegEx() {
     return begRegEx;
-  }
-
-  public List<Integer> getGroupsToMatch() {
-    return groupsToMatch;
   }
 
   public String getEndRegEx() {
@@ -106,15 +43,15 @@ public class ElementContext extends LanguageContext implements GrammarElement {
   @Override
   public String toString() {
     return "ElementContext{" +
-      "completionList=" + completionList +
+      "begRegEx='" + begRegEx + '\'' +
+      ", endRegEx='" + endRegEx + '\'' +
+      ", line='" + line + '\'' +
+      ", position=" + position +
+      ", capturedGroups=" + capturedGroups +
       ", isMatched=" + isMatched +
       ", isInScope=" + isInScope +
-      ", capturedGroups=" + capturedGroups +
       ", matchStart=" + matchStart +
       ", matchEnd=" + matchEnd +
-      ", begRegEx='" + begRegEx + '\'' +
-      ", endRegEx='" + endRegEx + '\'' +
-      ", groupsToMatch=" + groupsToMatch +
       '}';
   }
 }
